+++ tycho-integration-test/src/stream_processor/protocol_stream_processor.rs
@@ -17,6 +17,7 @@ use tycho_simulation::{
             aerodrome_slipstreams::state::AerodromeSlipstreamsState,
             cowamm::state::CowAMMState,
             ekubo::state::EkuboState,
+            ekubo_v3::state::EkuboV3State,
             erc4626::state::ERC4626State,
             filters::{balancer_v2_pool_filter, erc4626_filter, fluid_v1_paused_pools_filter},
             fluid::FluidV1,
@@ -161,6 +162,7 @@ impl ProtocolStreamProcessor {
                 "rocketpool".to_string(),
                 "erc4626".to_string(),
                 "cowamm".to_string(),
+                "ekubo_v3".to_string(),
             ],
             Chain::Base => vec![
                 "uniswap_v2".to_string(),
@@ -224,6 +226,9 @@ impl ProtocolStreamProcessor {
             }
             "ekubo_v2" => {
                 stream = stream.exchange::<EkuboState>("ekubo_v2", tvl_filter.clone(), None);
+            }
+            "ekubo_v3" => {
+                stream = stream.exchange::<EkuboV3State>("ekubo_v3", tvl_filter.clone(), None);
             }
             "vm:curve" => {
                 stream = stream.exchange::<EVMPoolState<PreCachedDB>>(